+++ src/solvers/trust_regions.jl
@@ -495,7 +495,7 @@ function trust_regions!(
     augmentation_threshold::R=0.75,
     augmentation_factor::R=2.0,
     sub_kwargs=(;),
-    sub_objective=decorate_objective!(M, TrustRegionModelObjective(mho), sub_kwargs...),
+    sub_objective=decorate_objective!(M, TrustRegionModelObjective(mho); sub_kwargs...),
     sub_problem=DefaultManoptProblem(TangentSpace(M, p), sub_objective),
     sub_stopping_criterion::StoppingCriterion=StopAfterIteration(manifold_dimension(M)) |
                                               StopWhenResidualIsReducedByFactorOrPower(;
@@ -513,8 +513,8 @@ function trust_regions!(
             trust_region_radius,
             randomize=randomize,
             (project!)=project!,
-            sub_kwargs...,
             stopping_criterion=sub_stopping_criterion,
+            sub_kwargs...,
         );
         sub_kwargs...,
     ),